% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catpull.R
\name{catpull}
\alias{catpull}
\title{Pull and clean set-level catch data}
\usage{
catpull(
  spec = NULL,
  minyear = 1995,
  maxyear = substr(Sys.Date(), 1, 4),
  lgroup = 5,
  username = oracle.username,
  password = oracle.password,
  dsn = oracle.dsn
)
}
\arguments{
\item{spec}{A species code.}

\item{minyear}{A minimum year. Default is 1995, when diet data collection began}

\item{maxyear}{A maximum year. Default is the present year.}

\item{username}{Oracle username. Default is oracle.username for those with an .Rprofile.}

\item{password}{Oracle password. Default is oracle.password for those with an .Rprofile.}

\item{dsn}{Oracle dsn. Defauls is oracle.dsn for those with an .Rprofile.}
}
\value{
A dataframe of length observations for a given species with length groups defined.
}
\description{
Pull and clean set-level catch data for a given species from the survey database.
}
\examples{
x <- catpull(spec=10)

}
