% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datapulls.R
\name{datapull}
\alias{datapull}
\title{Pull and clean diet data from .csv file}
\usage{
datapull(file = NULL, lgroup = 5)
}
\arguments{
\item{file}{A file name address.}

\item{lgroup}{Length groups used for selecting individual fish to record stomach contents. Default is 5 because, for the most species, the requirement for stomach sampling is 1 per 5cm.}
}
\value{
A dataframe of data without missing prey weights and with length groups specified.
}
\description{
Pull and clean diet data from a .csv file provided by the Food Habits Program.
}
\examples{
x <- datapull(file="C:/USER/filename.csv")

}
