% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detpull.R
\name{detpull}
\alias{detpull}
\title{Pull and clean length data}
\usage{
detpull(
  spec = NULL,
  minyear = 1995,
  maxyear = substr(Sys.Date(), 1, 4),
  lgroup = 5,
  username = oracle.username,
  password = oracle.password,
  dsn = oracle.dsn
)
}
\arguments{
\item{spec}{A species code.}

\item{minyear}{A minimum year. Default is 1995, when diet data collection began}

\item{maxyear}{A maximum year. Default is the present year.}

\item{lgroup}{Length groups used for selecting individual fish to record stomach contents. Default is 5 because, for the most species, the requirement for stomach sampling is 1 per 5cm.}

\item{username}{Oracle username. Default is oracle.username for those with an .Rprofile.}

\item{password}{Oracle password. Default is oracle.password for those with an .Rprofile.}

\item{dsn}{Oracle dsn. Defauls is oracle.dsn for those with an .Rprofile.}
}
\value{
A dataframe of length observations for a given species with length groups defined.
}
\description{
Pull and clean detailed length data for a given species from the survey database.
}
\examples{
x <- detpull(spec=10)

}
